chore(actions): refactor actions into separate files
